requirements.txt
azure-identity>=1.19.0,<2.0.0
Flask>=3.1.0,<4.0.0
gunicorn>=23.0.0,<24.0.0
pyodbc>=5.2.0,<6.0.0
